pyproject.toml
[build-system]
requires = ["setuptools >= 65", "wheel >= 0.38"]
build-backend = "setuptools.build_meta"

[project]
name = "kmake"
authors = [
    { name = "Antmicro" }
]
description = "Simple utility that automates handling of KiCad projects."
requires-python = ">=3.7"
version = "7.0.0"
license = {file = "LICENSE"}
dependencies = [
    "coloredlogs",
    "kiutils @ git+https://github.com/antmicro/kiutils.git@397f30e9aa464d82db6b18664728a9539b2675cd",
    "natsort==8.3.1",
    "termcolor",
    "Pillow",
    "pyxdg",
    "kicad_netlist_reader",
    "gitpython",
    "typing_extensions"
]

[tool.setuptools]
package-dir = { "kmake" = "src" }

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "mypy",
    "ruff"
]

[project.scripts]
kmake = 'kmake.kmake:main'

[tool.ruff.lint]
select = ["E", "F", "N", "B", "A", "RET"]

[tool.ruff]
line-length = 120
target-version = "py38"
exclude = [
	"docs",
	"src/commands/__init__.py",
	"src/ext_modules/__init__.py"
]

[tool.black]
line-length = 120

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = [
  "./src"
]
